WebApp/backend/requirements.txt
waitress
pandas
flask
scikit-learn==1.0.2
numpy
tensorflow-cpu
nltk